Create fitbit and applehealth min/max/avg heart rate data providers
