+++ src/helpers/daily-data-providers/index.ts
@@ -1,7 +1,9 @@
 ﻿export { default as appleHealthFlightsClimbedDataProvider } from "./apple-health-flights-climbed"
 export { default as appleHealthHrvDataProvider } from "./apple-health-hrv"
 export { default as appleHealthHeartRateRangeDataProvider } from "./apple-health-heart-rate-range"
-export { default as appleHealthMaxHeartRateDataProvider } from "./apple-health-max-heart-rate"
+export { maxHeartRate as appleHealthMaxHeartRateDataProvider } from "./apple-health-heart-rate"
+export { minHeartRate as appleHealthMinHeartRateDataProvider } from "./apple-health-heart-rate"
+export { averageHeartRate as appleHealthAverageHeartRateDataProvider } from "./apple-health-heart-rate"
 export { default as appleHealthRestingHeartRateDataProvider } from "./apple-health-resting-heart-rate"
 export { asleepTime as appleHealthSleepDataProvider } from "./apple-health-sleep"
 export { asleepRemTime as appleHealthSleepRemDataProvider } from "./apple-health-sleep"
